refactor(modules): move tests into separate file
